tests/submission/Ms2ger/dynamic-markup-insertion/innerhtml-02.html

<!DOCTYPE html>
<title>innerHTML in HTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-innerhtml">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var el = document.createElement("div");
  el.appendChild(document.createElement("xmp"))
    .appendChild(document.createElement("span"))
    .appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "<xmp><span>&lt;<\/span><\/xmp>");
})
test(function() {
  var el = document.createElement("xmp");
  el.appendChild(document.createElement("span"))
    .appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "<span>&lt;<\/span>");
})
test(function() {
  var el = document.createElement("xmp");
  el.appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "<");
})
test(function() {
  var el = document.createElement("div");
  el.appendChild(document.createElement("br"));
  assert_equals(el.innerHTML, "<br>");
})
</script>
